Rawspeed: TiffEntry::TiffEntry(): avoid unintentional integer overflow
overflow_before_widen: Potentially overflowing expression
    _count << RawSpeed::datashifts[_type]
with type RawSpeed::uint32 (32 bits, unsigned)
is evaluated using 32-bit arithmetic,
and then used in a context that expects an expression
of type uint64 (64 bits, unsigned).

Coverity CID 1357657

+++ src/external/rawspeed/RawSpeed/TiffEntry.cpp
@@ -85,7 +85,7 @@ TiffEntry::TiffEntry(TiffTag _tag, TiffDataType _type, uint32 _count, const ucha
   type = _type;
   count = _count;
   data_offset = -1; // Set nonsense value in case someone tries to use it
-  bytesize = _count << datashifts[_type];
+  bytesize = (size_t)_count << (size_t)datashifts[_type];
   if (NULL == _data) {
     own_data = new uchar8[bytesize];
     memset(own_data,0,bytesize);
